feat(ui): hydrate list pages on the server
